Move logger utilities to logger
